+++ QnA-Maker-KB-recovery/LICENSE
