+++ QnA-Maker-KB-recovery/requirement.txt
@@ -1,0 +1,3 @@
+azure-search-documents
+azure-core
+pandas
